\name{PKmodelFinder-package}
\alias{PKmodelFinder-package}
\alias{PKmodelFinder}
\docType{package}
\title{
Software for Pharmacokinetic modeling.
}
\description{
PKmodelFinder is the software for the exploratory analysis for pharmacometrics. 
It helps pharmacometricians explore their data, find adequate PK model,  
and diagnosis their fitted model. This software is build on R and provides
 a user-friendly graphical interface for the graphics and statistical analysis 
 without any other software for PK modeling.  Therefore, it is very helpful
  for many clinicians  fit their model and diagnosis their model by clicking mouse button.
}

\author{Eun-Kyung Lee,Gyejeong Noh, Hyeong-Seok Lim}


\keyword{ utilities }

\examples{
PKmodelFinder()
}
